refactor z pietrkiem
