+++ src/globals.js
@@ -1,0 +1,10 @@
+export const API_KEY = "f8a6e778f521e80470d9d262";
+
+const currencyOne = document.querySelector("#currency-one");
+const currencyTwo = document.querySelector("#currency-two");
+const amountOne = document.querySelector("#amount-one");
+const amountTwo = document.querySelector("#amount-two");
+const swapButton = document.querySelector("#swap");
+const rateInfo = document.querySelector("#rate");
+
+export { currencyOne,currencyTwo,amountOne,amountTwo,swapButton,rateInfo }
