+++ src/index.js
@@ -1,0 +1,16 @@
+// @ts-nocheck
+import { populateOptions } from './modules/dom-manipulator.module.js'
+
+// Event listeners
+
+
+
+
+// Add options to select
+
+
+export const initiExchangeRateCaluclator = () => {
+    populateOptions()
+}
+
+
